make some client methods public (#731)
* make some client methods public

* fix callers

* remove useless condition

* move observers models to reusable package
